@@ -1,19 +1,27 @@
 <template>
   <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js App"/>
+    <img alt="Vue logo" src="./assets/logo.png" />
+    <HelloWorld msg="Welcome to Your Vue 3 beta App" />
+    <Calc></Calc>
+    <FullName></FullName>
+    <Counter></Counter>
   </div>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld.vue'
+import HelloWorld from "./components/HelloWorld.vue";
+import FullName from "./components/FullName.vue";
+import Calc from "./components/Calc.vue";
+import Counter from "./components/Counter.vue";
 
 export default {
-  name: 'App',
   components: {
-    HelloWorld
-  }
-}
+    HelloWorld,
+    FullName,
+    Calc,
+    Counter,
+  },
+};
 </script>
 
 <style>
